Refactor: move tree logic to src/tree.ts and npm logic to src/npm.ts
